+++ packages/xstate-store/test/UseActor.vue
@@ -1,0 +1,28 @@
+<template>
+  <div data-testid="count">{{ snapshot.context.count }}</div>
+  <button data-testid="increment" @click="send({ type: 'inc' })">
+    Increment
+  </button>
+</template>
+
+<script lang="ts">
+import { defineComponent } from 'vue';
+import { useActor } from '@xstate/vue';
+
+import { fromStore } from '../src/index.ts';
+
+export default defineComponent({
+  emits: ['rerender'],
+  setup() {
+    const { snapshot, send } = useActor(fromStore({
+      count: 0
+    }, {
+      inc: (ctx) => ({ count: ctx.count + 1 })
+    }))
+
+    snapshot.value.context.count satisfies number;
+
+    return { snapshot, send };
+  }
+});
+</script>